Restructure structure to allow for electron
The `frontend` and `backend` folders were moved under the new `app`
folder to be bundled together
