+++ app/backend/lib/consts.js
@@ -7,14 +7,11 @@
 export const UPDATE_CHECK_INTERVAL = 1000 * 60
 
 // Setting file locations
-export const SETTINGS_FILE = './settings.json'
-export const DEFAULT_SETTINGS_FILE = './settings.default.json'
+export const SETTINGS_FILE = `${__dirname}/../settings.json`
+export const DEFAULT_SETTINGS_FILE = `${__dirname}/../settings.default.json`
 
 // Temporary files directory
 export const TEMP_DIR = './tmp'
-
-// ShabadOS database path
-export const DB_PATH = './database.sqlite'
 
 // Max Search results to return in one go
 export const MAX_RESULTS = 20
