+++ app/frontend/public/manifest.json
