+++ app/frontend/src/components/Controller/Navigator.css
