+++ app/frontend/src/components/Display.css
