+++ app/frontend/src/components/Display.js
