+++ app/frontend/src/fonts/noto-sans/NotoSansGurmukhi-Regular.ttf
