--- backend/package.json
@@ -1,46 +1,0 @@
-{
-  "name": "@shabados/backend",
-  "version": "1.0.0",
-  "description": "",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node -r @std/esm server.mjs",
-    "dev": "nodemon -e mjs -L --exec npm start | bunyan"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/ShabadOS/core.git"
-  },
-  "author": "ShabadOS Team <shabados.com@gmail.com>",
-  "license": "GPL-3.0",
-  "bugs": {
-    "url": "https://github.com/ShabadOS/core/issues"
-  },
-  "homepage": "https://github.com/ShabadOS/core#readme",
-  "dependencies": {
-    "@shabados/database": "^3.0.0",
-    "@std/esm": "^0.25.5",
-    "body-parser": "^1.18.2",
-    "bunyan": "^1.8.12",
-    "compression": "^1.7.2",
-    "download-github-release": "^0.3.2",
-    "express": "^4.16.3",
-    "fs-extra": "^5.0.0",
-    "helmet": "^3.12.0",
-    "knex": "^0.14.4",
-    "objection": "^1.0.1",
-    "progress": "^2.0.0",
-    "semver": "^5.5.0",
-    "update-notifier": "^2.4.0",
-    "ws": "^5.1.0"
-  },
-  "devDependencies": {
-    "eslint": "^4.19.1",
-    "eslint-config-airbnb": "^16.1.0",
-    "eslint-plugin-import": "^2.9.0",
-    "eslint-plugin-jsx-a11y": "^6.0.3",
-    "eslint-plugin-react": "^7.7.0",
-    "nodemon": "^1.17.2"
-  },
-  "@std/esm": "cjs"
-}
